pyproject.toml
[tool.poetry]
name = "vyp"
version = "0.1.0"
description = ""
authors = ["Ivan Feofanov <feofanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
eth-ape = "^0.7.10"
ape-hardhat = "^0.7.0"
ape-vyper = "^0.7.0"
vyper = "^0.3.10"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
faker = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
